requirements.txt
Pillow==9.4.0
PyPDF2==3.0.1
gTTS==2.3.0
requests==2.31.0
tqdm==4.65.0
